+++ packages/core/src/table/hooks/use_table_ref.tsx
@@ -10,9 +10,13 @@ export const useTableRef = () => {
     const listener = (table: Table<any>) => {
       tableRef.current = table;
     };
-    context.updateListeners.table.add(listener);
+    const listenerEntry: any = {
+      callback: listener,
+      dependency: { type: "table" },
+    };
+    context.updateListeners.table.add(listenerEntry);
     return () => {
-      context.updateListeners.table.delete(listener);
+      context.updateListeners.table.delete(listenerEntry);
     };
   }, [context]);