+++ packages/skin-mui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rttui/skin-mui
 
+## 1.0.30
+
+### Patch Changes
+
+- Updated dependencies
+  - @rttui/core@1.0.29
+
 ## 1.0.29
 
 ### Patch Changes